+++ Rdmp.Core/Icons/IconProvision/CatalogueIcons.Designer.cs
@@ -2035,18 +2035,6 @@ namespace Rdmp.Core.Icons.IconProvision
         /// <summary>
         ///   Looks up a localized resource of type Image.
         /// </summary>
-        public static Byte[] CommittedCohortIdentificationNode
-        {
-            get
-            {
-                object obj = ResourceManager.GetObject("CommittedCohortIdentificationNode", resourceCulture);
-                return ((Byte[])(obj));
-            }
-        }
-
-        /// <summary>
-        ///   Looks up a localized resource of type Image.
-        /// </summary>
         public static Byte[] ProjectsNode
         {
             get