+++ Rdmp.Core/Icons/IconProvision/RDMPConcept.cs
@@ -207,7 +207,6 @@ public enum RDMPConcept
     LoadMetadataCatalogueLinkage,
     Setting,
     TicketingSystemReleaseStatus,
-    CommittedCohortIdentificationNode,
     AllDatasetsNode,
     RegexRedaction,
     RegexRedactionConfiguration,